V2-2: Clean up unit testing
