+++ mtag_sdk/build.gradle
@@ -8,6 +8,7 @@ android {
         unitTests.returnDefaultValues = true
     }
 
+
     defaultConfig {
         minSdkVersion 23
         targetSdkVersion 28
@@ -32,6 +33,7 @@ dependencies {
 
     implementation 'com.android.support:appcompat-v7:28.0.0'
     testImplementation 'junit:junit:4.12'
+    testImplementation 'org.json:json:20180130'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
     implementation 'com.loopj.android:android-async-http:1.4.9'
